Rename or1k-elf-loader to elf-loader
There's nothing or1k specific to this 'core' anymore, so let's
use a more common name for it.
