+++ cores/elf-loader/elf-loader.c
