+++ cores/elf-loader/elf-loader.core
@@ -1,0 +1,12 @@
+CAPI=1
+[main]
+description = "Generic ELF loader"
+
+[vpi]
+name = elf_loader
+src_files = elf-loader.c vpi_wrapper.c
+include_files = elf-loader.h
+libs = -lelf
+
+[plusargs]
+elf-load = file ELF file to preload to memory
