+++ cores/elf-loader/elf-loader.h
