--- cores/or1k-elf-loader/or1k-elf-loader.core
@@ -1,12 +1,0 @@
-CAPI=1
-[main]
-description = "OR1K elf loader"
-
-[vpi]
-name = or1k_elf_loader
-src_files = or1k-elf-loader.c vpi_wrapper.c
-include_files = or1k-elf-loader.h
-libs = -lelf
-
-[plusargs]
-or1k-elf-load = file ELF file to preload to memory
