+++ systems/de1/de1.core
@@ -6,7 +6,7 @@ depend =
  adv_debug_sys
  or1200
  uart16550
- or1k-elf-loader
+ elf-loader
  vlog_tb_utils
  jtag_vpi
  wiredelay
