+++ systems/generic/generic.core
@@ -6,7 +6,7 @@ depend =
  jtag_tap
  jtag_vpi
  or1200
- or1k-elf-loader
+ elf-loader
  ram_wb
  uart16550
  vlog_tb_utils
