+++ systems/or1200-generic/bench/verilog/orpsoc_tb.v
@@ -15,13 +15,13 @@ module orpsoc_tb;
    reg [1023:0] elf_file;
 
    initial begin
-      if($value$plusargs("or1k_elf_load=%s", elf_file)) begin
-	 $or1k_elf_load_file(elf_file);
+      if($value$plusargs("elf_load=%s", elf_file)) begin
+	 $elf_load_file(elf_file);
 	 
-	 mem_words = $or1k_elf_get_size/4;
+	 mem_words = $elf_get_size/4;
 	 $display("Loading %d words", mem_words);
 	 for(i=0; i < mem_words; i = i+1)
-	   orpsoc_tb.dut.wb_bfm_memory0.mem[i] = $or1k_elf_read_32(i*4);
+	   orpsoc_tb.dut.wb_bfm_memory0.mem[i] = $elf_read_32(i*4);
       end else
 	$display("No ELF file specified");
